REFACTOR: Move model code into model folder
