+++ packages/documents/src/index.ts
@@ -1,10 +1,10 @@
 export { createBinder } from "./binder";
 export type { Binder } from "./binder";
-export type { MorphismCell, ObjectCell } from "./cell";
-export type { ModelDocument } from "./model-document";
+export type { MorphismCell, ObjectCell } from "./model/cell";
+export type { ModelDocument } from "./model/document";
+export { notebookFromModel } from "./model/notebook";
+export type { Notebook } from "./model/notebook";
 export type { NotebookDocument } from "./notebook-document";
-export type { Notebook } from "./notebook";
-export { notebookFromModel } from "./notebook";
 export type { RichTextCell } from "./rich-text";
 export { defineMorphism, defineObject, defineShape, RichText } from "./shape";
 export type { MorphismEndpoint, MorphismEndpoints, MorphismType, ObjectType, Shape } from "./shape";
